更新圖片顯示

+++ src/views/PlanDetail.vue
@@ -5,16 +5,21 @@ const router = useRouter()
 const { planID } = route.params
 console.log(planID)
 const planDelete = async (id) => {
-    const API_URL = `https://localhost:7127/api/Plans/${id}`
-    const res = await fetch(API_URL, {
-        method: 'DELETE',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-    })
-    router.push({
-        path: `/plan/chart`,
-    })
+   
+const isDelete = confirm('確定刪除資料?')
+if(isDelete){
+  
+const API_URL = `https://localhost:7127/api/Plans/${id}`
+const res = await fetch(API_URL, {
+    method: 'DELETE',
+    headers: {
+        'Content-Type': 'application/json'
+    },
+})
+router.push({
+    path: `/plan/chart`,
+})
+}
 }
 
 const changeDiet = () => {
@@ -30,7 +35,7 @@ const changeSport = () => {
     })
 }
 
-const planBack =() =>{
+const planBack = () => {
     router.push({
         path: `/plan/chart`,
     })
